backend-django/requirement.txt
Django==4.2
djangorestframework==3.14
mysqlclient==2.1.1
python-decouple==3.8